+++ ChangeLog.md
@@ -5,6 +5,9 @@
   * Add 'SetState' Cottle function.  Details on how to use this are in the SpeechResponder documentation
   * Allow speech responder to not speak when subtitles are written
   * Retry companion API profile call if returned information is outdated.  This should help to avoid situations where outfitting and market data is out-of-date
+  * Add event when being scanned for cargo by an NPC
+  * Add event when being attacked by an NPC
+  * Add event when being interdicted by an NPC
 
 #2.1.0-b1
   * Fix material name for Cadmium